Rename reducer name
